unified ChatModelListenerIT
