+++ langchain4j-azure-open-ai/src/test/java/dev/langchain4j/model/azure/AzureOpenAiChatModelListenerIT.java
@@ -1,0 +1,47 @@
+package dev.langchain4j.model.azure;
+
+import com.azure.core.exception.ClientAuthenticationException;
+import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.chat.ChatModelListenerIT;
+import dev.langchain4j.model.chat.listener.ChatModelListener;
+
+import static java.util.Collections.singletonList;
+
+class AzureOpenAiChatModelListenerIT extends ChatModelListenerIT {
+
+    @Override
+    protected ChatLanguageModel createModel(ChatModelListener listener) {
+        return AzureOpenAiChatModel.builder()
+                .endpoint(System.getenv("AZURE_OPENAI_ENDPOINT"))
+                .apiKey(System.getenv("AZURE_OPENAI_KEY"))
+                .deploymentName(modelName())
+                .temperature(temperature())
+                .topP(topP())
+                .maxTokens(maxTokens())
+                .logRequestsAndResponses(true)
+                .listeners(singletonList(listener))
+                .build();
+    }
+
+    @Override
+    protected String modelName() {
+        return "gpt-4o-mini";
+    }
+
+    @Override
+    protected ChatLanguageModel createFailingModel(ChatModelListener listener) {
+        return AzureOpenAiChatModel.builder()
+                .endpoint(System.getenv("AZURE_OPENAI_ENDPOINT"))
+                .apiKey("banana")
+                .deploymentName(modelName())
+                .maxRetries(1)
+                .logRequestsAndResponses(true)
+                .listeners(singletonList(listener))
+                .build();
+    }
+
+    @Override
+    protected Class<?> expectedExceptionClass() {
+        return ClientAuthenticationException.class;
+    }
+}
