+++ langchain4j-google-ai-gemini/src/test/java/dev/langchain4j/model/googleai/GoogleAiGeminiChatModelListenerIT.java
@@ -1,0 +1,49 @@
+package dev.langchain4j.model.googleai;
+
+import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.chat.ChatModelListenerIT;
+import dev.langchain4j.model.chat.listener.ChatModelListener;
+
+import static java.util.Collections.singletonList;
+
+class GoogleAiGeminiChatModelListenerIT extends ChatModelListenerIT {
+
+    @Override
+    protected ChatLanguageModel createModel(ChatModelListener listener) {
+        return GoogleAiGeminiChatModel.builder()
+                .apiKey(System.getenv("GOOGLE_AI_GEMINI_API_KEY"))
+                .modelName(modelName())
+                .temperature(temperature())
+                .topP(topP())
+                .maxOutputTokens(maxTokens())
+                .listeners(singletonList(listener))
+                .logRequestsAndResponses(true)
+                .build();
+    }
+
+    @Override
+    protected String modelName() {
+        return "gemini-1.5-flash";
+    }
+
+    @Override
+    protected boolean assertResponseId() {
+        return false;
+    }
+
+    @Override
+    protected ChatLanguageModel createFailingModel(ChatModelListener listener) {
+        return GoogleAiGeminiChatModel.builder()
+                .apiKey("banana")
+                .modelName(modelName())
+                .maxRetries(1)
+                .listeners(singletonList(listener))
+                .logRequestsAndResponses(true)
+                .build();
+    }
+
+    @Override
+    protected Class<?> expectedExceptionClass() {
+        return RuntimeException.class;
+    }
+}
