+++ langchain4j-open-ai/src/test/java/dev/langchain4j/model/openai/OpenAiChatModelIT.java
@@ -1,6 +1,5 @@
 package dev.langchain4j.model.openai;
 
-import dev.ai4j.openai4j.OpenAiHttpException;
 import dev.langchain4j.agent.tool.ToolExecutionRequest;
 import dev.langchain4j.agent.tool.ToolSpecification;
 import dev.langchain4j.data.message.AiMessage;
@@ -12,19 +11,12 @@ import dev.langchain4j.data.message.UserMessage;
 import dev.langchain4j.internal.Json;
 import dev.langchain4j.model.Tokenizer;
 import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.chat.listener.ChatModelErrorContext;
-import dev.langchain4j.model.chat.listener.ChatModelListener;
-import dev.langchain4j.model.chat.listener.ChatModelRequest;
-import dev.langchain4j.model.chat.listener.ChatModelRequestContext;
-import dev.langchain4j.model.chat.listener.ChatModelResponse;
-import dev.langchain4j.model.chat.listener.ChatModelResponseContext;
 import dev.langchain4j.model.output.Response;
 import dev.langchain4j.model.output.TokenUsage;
 import org.junit.jupiter.api.Test;
 
 import java.util.Base64;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static dev.langchain4j.agent.tool.JsonSchemaProperty.INTEGER;
 import static dev.langchain4j.data.message.ToolExecutionResultMessage.from;
@@ -37,8 +29,6 @@ import static dev.langchain4j.model.output.FinishReason.TOOL_EXECUTION;
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Fail.fail;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class OpenAiChatModelIT {
 
@@ -447,131 +437,5 @@ class OpenAiChatModelIT {
 
         // then
         assertThat(tokenCount).isEqualTo(42);
-    }
-
-    @Test
-    void should_listen_request_and_response() {
-
-        // given
-        AtomicReference<ChatModelRequest> requestReference = new AtomicReference<>();
-        AtomicReference<ChatModelResponse> responseReference = new AtomicReference<>();
-
-        ChatModelListener listener = new ChatModelListener() {
-
-            @Override
-            public void onRequest(ChatModelRequestContext requestContext) {
-                requestReference.set(requestContext.request());
-                requestContext.attributes().put("id", "12345");
-            }
-
-            @Override
-            public void onResponse(ChatModelResponseContext responseContext) {
-                responseReference.set(responseContext.response());
-                assertThat(responseContext.request()).isSameAs(requestReference.get());
-                assertThat(responseContext.attributes().get("id")).isEqualTo("12345");
-            }
-
-            @Override
-            public void onError(ChatModelErrorContext errorContext) {
-                fail("onError() must not be called");
-            }
-        };
-
-        OpenAiChatModelName modelName = GPT_4_O_MINI;
-        double temperature = 0.7;
-        double topP = 1.0;
-        int maxTokens = 7;
-
-        OpenAiChatModel model = OpenAiChatModel.builder()
-                .baseUrl(System.getenv("OPENAI_BASE_URL"))
-                .apiKey(System.getenv("OPENAI_API_KEY"))
-                .organizationId(System.getenv("OPENAI_ORGANIZATION_ID"))
-                .modelName(modelName)
-                .temperature(temperature)
-                .topP(topP)
-                .maxTokens(maxTokens)
-                .logRequests(true)
-                .logResponses(true)
-                .listeners(singletonList(listener))
-                .build();
-
-        UserMessage userMessage = UserMessage.from("hello");
-
-        ToolSpecification toolSpecification = ToolSpecification.builder()
-                .name("add")
-                .addParameter("a", INTEGER)
-                .addParameter("b", INTEGER)
-                .build();
-
-        // when
-        AiMessage aiMessage = model.generate(singletonList(userMessage), singletonList(toolSpecification)).content();
-
-        // then
-        ChatModelRequest request = requestReference.get();
-        assertThat(request.model()).isEqualTo(modelName.toString());
-        assertThat(request.temperature()).isEqualTo(temperature);
-        assertThat(request.topP()).isEqualTo(topP);
-        assertThat(request.maxTokens()).isEqualTo(maxTokens);
-        assertThat(request.messages()).containsExactly(userMessage);
-        assertThat(request.toolSpecifications()).containsExactly(toolSpecification);
-
-        ChatModelResponse response = responseReference.get();
-        assertThat(response.id()).isNotBlank();
-        assertThat(response.model()).isNotBlank();
-        assertThat(response.tokenUsage().inputTokenCount()).isGreaterThan(0);
-        assertThat(response.tokenUsage().outputTokenCount()).isGreaterThan(0);
-        assertThat(response.tokenUsage().totalTokenCount()).isGreaterThan(0);
-        assertThat(response.finishReason()).isNotNull();
-        assertThat(response.aiMessage()).isEqualTo(aiMessage);
-    }
-
-    @Test
-    void should_listen_error() {
-
-        // given
-        String wrongApiKey = "banana";
-
-        AtomicReference<ChatModelRequest> requestReference = new AtomicReference<>();
-        AtomicReference<Throwable> errorReference = new AtomicReference<>();
-
-        ChatModelListener listener = new ChatModelListener() {
-
-            @Override
-            public void onRequest(ChatModelRequestContext requestContext) {
-                requestReference.set(requestContext.request());
-                requestContext.attributes().put("id", "12345");
-            }
-
-            @Override
-            public void onResponse(ChatModelResponseContext responseContext) {
-                fail("onResponse() must not be called");
-            }
-
-            @Override
-            public void onError(ChatModelErrorContext errorContext) {
-                errorReference.set(errorContext.error());
-                assertThat(errorContext.request()).isSameAs(requestReference.get());
-                assertThat(errorContext.partialResponse()).isNull();
-                assertThat(errorContext.attributes().get("id")).isEqualTo("12345");
-            }
-        };
-
-        OpenAiChatModel model = OpenAiChatModel.builder()
-                .apiKey(wrongApiKey)
-                .maxRetries(0)
-                .logRequests(true)
-                .logResponses(true)
-                .listeners(singletonList(listener))
-                .build();
-
-        String userMessage = "this message will fail";
-
-        // when
-        assertThrows(RuntimeException.class, () -> model.generate(userMessage));
-
-        // then
-        Throwable throwable = errorReference.get();
-        assertThat(throwable).isExactlyInstanceOf(OpenAiHttpException.class);
-        assertThat(throwable).hasMessageContaining("Incorrect API key provided");
     }
 }
