+++ langchain4j-open-ai/src/test/java/dev/langchain4j/model/openai/OpenAiChatModelListenerIT.java
@@ -1,0 +1,47 @@
+package dev.langchain4j.model.openai;
+
+import dev.ai4j.openai4j.OpenAiHttpException;
+import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.chat.ChatModelListenerIT;
+import dev.langchain4j.model.chat.listener.ChatModelListener;
+
+import static dev.langchain4j.model.openai.OpenAiChatModelName.GPT_4_O_MINI;
+import static java.util.Collections.singletonList;
+
+class OpenAiChatModelListenerIT extends ChatModelListenerIT {
+
+    @Override
+    protected ChatLanguageModel createModel(ChatModelListener listener) {
+        return OpenAiChatModel.builder()
+                .baseUrl(System.getenv("OPENAI_BASE_URL"))
+                .apiKey(System.getenv("OPENAI_API_KEY"))
+                .organizationId(System.getenv("OPENAI_ORGANIZATION_ID"))
+                .modelName(modelName())
+                .temperature(temperature())
+                .topP(topP())
+                .maxTokens(maxTokens())
+                .logRequests(true)
+                .logResponses(true)
+                .listeners(singletonList(listener))
+                .build();
+    }
+
+    @Override
+    protected String modelName() {
+        return GPT_4_O_MINI.toString();
+    }
+
+    @Override
+    protected ChatLanguageModel createFailingModel(ChatModelListener listener) {
+        return OpenAiChatModel.builder()
+                .apiKey("banana")
+                .maxRetries(1)
+                .listeners(singletonList(listener))
+                .build();
+    }
+
+    @Override
+    protected Class<?> expectedExceptionClass() {
+        return OpenAiHttpException.class;
+    }
+}
